ai_prayer/requirements.txt
aws-cdk-lib==2.201.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha==2.140.0a0
boto3
python-dotenv
elevenlabs